added basic ts setup
