+++ .gitignore
@@ -123,3 +123,4 @@ demo*
 usageReadme.txt
 
 *_flat*
+typechain/
