+++ deployments/localhost/SushiswapFactory.json
@@ -1,0 +1,23 @@
+{
+  "address": "0x8629A03B95c75bC27278122F5A00438439dAF628",
+  "abi": [],
+  "transactionHash": "0x1513d7f785693db06cfb92bbcf0c6fa6007433ecd517122481795e926bf9a85c",
+  "receipt": {
+    "to": null,
+    "from": "0xD8d8632Bb8C8b199e43faDf7205749dd34C4B8c9",
+    "contractAddress": "0x8629A03B95c75bC27278122F5A00438439dAF628",
+    "transactionIndex": 0,
+    "gasUsed": "67054",
+    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+    "blockHash": "0xdfceedc628db4b9bd4c0a62e28cc362f160c3f49bc5526f187a3d22c615c1963",
+    "transactionHash": "0x1513d7f785693db06cfb92bbcf0c6fa6007433ecd517122481795e926bf9a85c",
+    "logs": [],
+    "blockNumber": 15,
+    "cumulativeGasUsed": "67054",
+    "status": 1,
+    "byzantium": true
+  },
+  "args": [],
+  "bytecode": "0x6080604052348015600f57600080fd5b50603f80601d6000396000f3fe6080604052600080fdfea264697066735822122001d50454e5d7752b3bb678db17f142c36ddfa3fab0458472f12ba18c0667f1d364736f6c63430008000033",
+  "deployedBytecode": "0x6080604052600080fdfea264697066735822122001d50454e5d7752b3bb678db17f142c36ddfa3fab0458472f12ba18c0667f1d364736f6c63430008000033"
+}
