+++ deployments/localhost/UniswapFactory.json
@@ -1,0 +1,23 @@
+{
+  "address": "0xC4365928daDdC6514531324A997DBA590E68fe89",
+  "abi": [],
+  "transactionHash": "0x219eefca2d3a0911acefccbd4ccb509e9da4fbaf7d056556dc32078dd3a20d65",
+  "receipt": {
+    "to": null,
+    "from": "0xD8d8632Bb8C8b199e43faDf7205749dd34C4B8c9",
+    "contractAddress": "0xC4365928daDdC6514531324A997DBA590E68fe89",
+    "transactionIndex": 0,
+    "gasUsed": "67054",
+    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+    "blockHash": "0xfcee7796c6218493abb33a81e31011765ba7b241589fb60117a1a4040c407720",
+    "transactionHash": "0x219eefca2d3a0911acefccbd4ccb509e9da4fbaf7d056556dc32078dd3a20d65",
+    "logs": [],
+    "blockNumber": 14,
+    "cumulativeGasUsed": "67054",
+    "status": 1,
+    "byzantium": true
+  },
+  "args": [],
+  "bytecode": "0x6080604052348015600f57600080fd5b50603f80601d6000396000f3fe6080604052600080fdfea264697066735822122001d50454e5d7752b3bb678db17f142c36ddfa3fab0458472f12ba18c0667f1d364736f6c63430008000033",
+  "deployedBytecode": "0x6080604052600080fdfea264697066735822122001d50454e5d7752b3bb678db17f142c36ddfa3fab0458472f12ba18c0667f1d364736f6c63430008000033"
+}
