+++ hardhat.config.ts
@@ -1,0 +1,35 @@
+import { task } from "hardhat/config";
+import "@typechain/hardhat";
+import "@nomiclabs/hardhat-ethers";
+import "@nomiclabs/hardhat-waffle";
+
+// This is a sample Hardhat task. To learn how to create your own go to
+// https://hardhat.org/guides/create-task.html
+task("accounts", "Prints the list of accounts", async (args, hre) => {
+  const accounts = await hre.ethers.getSigners();
+
+  for (const account of accounts) {
+    console.log(await account.address);
+  }
+});
+
+// You need to export an object to set up your config
+// Go to https://hardhat.org/config/ to learn more
+
+/**
+ * @type import('hardhat/config').HardhatUserConfig
+ */
+export default {
+  solidity: "0.8.4",
+  networks: {
+    hardhat: {
+      chainId: 1337,
+      accounts: {
+        mnemonic: "eight fun oak spot hip pencil matter domain bright fiscal nurse easy",
+      },
+    },
+  },
+  mocha: {
+    timeout: 50000,
+  },
+};
